# docs/openapi.yaml
openapi: 3.1.0
info:
  title: MoneyApp API
  version: 1.0.0
  description: |
    Self-hosted budget and net-worth tracker.

    Conventions:
      * Session cookie auth. `user_id` is taken only from the session, never from input.
      * Money is always an object with integer minor units. No floats on the wire.
      * `null` means "not recorded". `0` means "recorded zero". Never interchangeable.
      * Errors are RFC 7807 application/problem+json.
      * Lists are cursor-paginated on (occurred_on, id).
  license:
    name: MIT
servers:
  - url: /api/v1
security:
  - sessionCookie: []
...
paths:
  /auth/login:
    post:
      tags: [auth]
      summary: Log in
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:    { type: string, format: email }
                password: { type: string, minLength: 8 }
      responses:
        "204":
          description: Session cookie set
          headers:
            Set-Cookie:
              schema: { type: string }
        "401": { $ref: "#/components/responses/Problem" }
        "429": { $ref: "#/components/responses/Problem" }
...
  /auth/me:
    get:
      tags: [auth]
      summary: Current user
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "401": { $ref: "#/components/responses/Problem" }
  /auth/password:
    post:
      tags: [auth]
      summary: Change the current user's password
      description: |
        Revokes every other session. While `must_change_password` is set this is the
        only endpoint the session may reach; everything else answers 403.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [current_password, new_password]
              properties:
                current_password: { type: string }
                new_password:     { type: string, minLength: 8 }
              additionalProperties: false
      responses:
        "204": { description: Changed }
        "401": { $ref: "#/components/responses/Problem" }
        "422": { $ref: "#/components/responses/Problem" }
  /auth/sessions:
    get:
      tags: [auth]
      summary: List the current user's live sessions
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Session" }
        "401": { $ref: "#/components/responses/Problem" }
  /auth/sessions/{id}:
    parameters:
      - name: id
        in: path
        required: true
        description: The session's id as returned by GET /auth/sessions.
        schema: { type: string }
    delete:
      tags: [auth]
      summary: Revoke one session
      responses:
        "204": { description: Revoked }
        "404": { $ref: "#/components/responses/Problem" }
...
  /budgets:
    get:
      tags: [budgets]
      summary: Budgets for a period
      parameters:
        - $ref: "#/components/parameters/PeriodQuery"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Budget" }
  /budgets/bulk:
    post:
      tags: [budgets]
      summary: Seed budgets across a period range
      description: Applies one planned figure per category to every month in the range.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [from_period, to_period, items]
              properties:
                from_period: { type: string, pattern: '^\d{4}-\d{2}$' }
                to_period:   { type: string, pattern: '^\d{4}-\d{2}$' }
                items:
                  type: array
                  items:
                    type: object
                    required: [category_id, planned]
                    properties:
                      category_id: { type: integer }
                      planned:     { $ref: "#/components/schemas/Money" }
      responses:
        "200":
          description: Applied
          content:
            application/json:
              schema:
                type: object
                properties:
                  periods_written: { type: integer }
                  rows_written:    { type: integer }
  /budgets/{category_id}/{period}:
    parameters:
      - name: category_id
        in: path
        required: true
        schema: { type: integer }
      - name: period
        in: path
        required: true
        schema: { type: string, pattern: '^\d{4}-\d{2}$' }
    put:
      tags: [budgets]
      summary: Set one category's plan for one month
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [planned]
              properties:
                planned: { $ref: "#/components/schemas/Money" }
              additionalProperties: false
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Budget" }
        "404": { $ref: "#/components/responses/Problem" }
        "422": { $ref: "#/components/responses/Problem" }
    delete:
      tags: [budgets]
      summary: Remove a plan
      description: |
        An absent plan is "not planned", which is a different fact from a plan of
        zero, so this is not the same as setting it to 0.
      responses:
        "204": { description: Deleted }
        "404": { $ref: "#/components/responses/Problem" }
...
  /reports/budget:
    get:
      tags: [reports]
      summary: Plan versus actual for a period
      parameters:
        - $ref: "#/components/parameters/PeriodQuery"
        - $ref: "#/components/parameters/Valuation"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BudgetReport" }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: moneyapp_session
  parameters:
...
    PeriodQuery:
      name: period
      in: query
      required: true
      schema: { type: string, pattern: '^\d{4}-\d{2}$' }
    Valuation:
      name: valuation
      in: query
      description: |
        contemporaneous values each period at its own rate.
        constant values everything at today's rate, reproducing the legacy spreadsheet.
      schema:
        type: string
        enum: [contemporaneous, constant]
        default: contemporaneous
...
  responses:
    Problem:
      description: RFC 7807 problem document
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
  schemas:
    Problem:
      type: object
      required: [type, title, status]
      properties:
        type:   { type: string }
        title:  { type: string }
        status: { type: integer }
        detail: { type: string }
        errors:
          type: array
          items:
            type: object
            properties:
              field:   { type: string }
              message: { type: string }
    Money:
      type: object
      required: [amount_minor, currency, exponent]
      properties:
        amount_minor:
          type: integer
          format: int64
          description: Value in minor units. HUF has exponent 0, so 1000 means 1000 HUF.
        currency: { type: string, minLength: 3, maxLength: 3 }
        exponent: { type: integer, minimum: 0, maximum: 4 }
...
    User:
      type: object
      properties:
        id:                      { type: integer }
        email:                   { type: string, format: email }
        display_name:            { type: string, nullable: true }
        role:                    { type: string, enum: [admin, user, readonly] }
        base_currency:           { type: string }
        timezone:                { type: string }
        fiscal_year_start_month: { type: integer, minimum: 1, maximum: 12 }
        must_change_password:
          type: boolean
          description: |
            While true, every endpoint except POST /auth/password answers 403.
    Session:
      type: object
      properties:
        id:
          type: string
          description: |
            The session's identifier. It is the SHA-256 of the token, not the token
            itself, so listing sessions never exposes a usable credential.
        user_agent: { type: string }
        created_at: { type: string, format: date-time }
        expires_at: { type: string, format: date-time }
        current:    { type: boolean, description: True for the session making the request. }
...
    Budget:
      type: object
      properties:
        category_id:   { type: integer }
        category_name: { type: string }
        period:        { type: string }
        planned:       { $ref: "#/components/schemas/Money" }
...
    BudgetReport:
      type: object
      properties:
        period:        { type: string }
        valuation:     { type: string }
        base_currency: { type: string }
        unconverted:
          type: integer
          description: |
            Transactions in this period with no rate for their date. They are
            excluded from the totals and reported rather than counted as zero.
        spend_total:
          oneOf:
            - $ref: "#/components/schemas/Money"
            - type: "null"
          description: null when the period recorded nothing at all.
        planned_total: { $ref: "#/components/schemas/Money" }
        income:
          oneOf:
            - $ref: "#/components/schemas/Money"
            - type: "null"
        diff:
          oneOf:
            - $ref: "#/components/schemas/Money"
            - type: "null"
        saved_percent:
          type: number
          nullable: true
          description: Ratio, unclamped. Negative when spend exceeds income; null when income is zero.
        possible_minimum: { $ref: "#/components/schemas/Money" }
        categories:
          type: array
          items:
            type: object
            properties:
              category_id: { type: integer }
              name:        { type: string }
              actual:
                oneOf:
                  - $ref: "#/components/schemas/Money"
                  - type: "null"
                description: null means the period has no recorded data, distinct from zero.
              icon:        { type: string, nullable: true }
              color:       { type: string, nullable: true }
              is_essential: { type: boolean }
              planned:
                oneOf:
                  - $ref: "#/components/schemas/Money"
                  - type: "null"
                description: null when no plan was recorded, which differs from a plan of zero.
              average:
                oneOf:
                  - $ref: "#/components/schemas/Money"
                  - type: "null"
                description: |
                  Mean over every recorded month, the workbook's column C. It holds
                  steady as the month on screen changes, and is null until there is
                  data to average — never 0.
              ratio:   { type: number, nullable: true }
              state:
                type: string
                enum: [severely_over, over, approaching, within, zero, not_recorded]
              state_label:
                type: string
                description: |
                  Accessible wording for the state, so colour is never the only
                  signal and every client says the same thing.
...